👌 IMPROVE:  separation of tests
